requirements.txt
fastapi==0.123.10
uvicorn==0.38.0
pydantic==2.10.6
supabase==2.15.3
python-dotenv==1.1.0
openai==1.82.0
httpx==0.27.2
aiohttp==3.12.13
loguru==0.7.3
apscheduler==3.11.2
requests==2.32.4
